requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
Django==2.2.10
django-bootstrap-modal-forms==1.5.0
django-countries==6.0
django-crispy-forms==1.9.0
django-widget-tweaks==1.4.8
idna==2.9
Pillow==7.0.0
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
stripe==2.43.0
urllib3==1.25.8
